timer:cancel: handle cancel of expired tasks
Handle cancel of already expired tasks.

If a one shot task already executed, or a repeating task halted, and
then a new task was added to its slot, the id of the original task would
cancel the new task:

Example of previous bad behavior:

auto t_1 = timer.in(1000, function);
// wait 1 or more seconds, or be canceled
auto t_2 = timer.in(1000, function);
timer.cancel(t_1); // might cancel t_2 if it is placed in t_1 slot

Now the ID is tied to the type, handler, its argument, and expiry. It is
computed instead of stored to save memory. The caveat is that if the
same task is added more than once with the same arguments, the ID
returned will cancel an arbitrary one of those tasks.

For example:

auto t_1 = timer.in(1000, function);
auto t_2 = timer.in(1000, function);
timer.cancel(t_1);

// an arbitrary one of either t_1 or t_2 is canceled (for .at() only
// if the time of creation is the same within the clock precision).
// In practice, this shouldn't be an issue, since they are likely
// identical (same type, handler, argument, and expiry), and the one
// canceled shouldn't matter. If you add it twice, and cancel one, the
// other still runs. The remaining id is still valid to cancel the
// remaining task (in this case t_2).